Real data integration
France data (18-12-2006) integrated and progresschart is also dynamic, there is a small bug I have not figured out yet but highlighted in the comments  of the second return statement in the heatmap file. This version serves as a backup as well because I will try to change progresschart to piechart since it makes more sense.